Fixed scale-agnostic image size comparison (#100)
Previously we would compare UIImage size properties, which are not in pixel units but scale-dependent points. This would cause issues if a test used reference images generated using .none fileNameOptions and ran tests on 2x or 3x simulators: they would fail early because of the size comparison because of the scale difference.

This commit changes the logic to use CGImageGet{Width,Height} instead, which returns the actual pixel size (akin to use size * scale) and is backwards compatible since the image comparison code actually uses CGImage pixel retrieving methods.